Complete DOP Framework

+++ DOPFramework/Assets/Dotity/Component/Component.cs
@@ -6,13 +6,13 @@ namespace Dotity
     public class Component : IComponent
     {
         #region Static Function
-        private static Dictionary<short, Stack<IComponent>> _listComponentsReuse = new Dictionary<short, Stack<IComponent>>();
+        private static Dictionary<int, Stack<IComponent>> _listComponentsReuse = new Dictionary<int, Stack<IComponent>>();
 
         public static T Create<T>(short keyComponent) where T : new()
         {
             return (_listComponentsReuse.ContainsKey(keyComponent) && _listComponentsReuse[keyComponent].Count > 0) ? (T)_listComponentsReuse[keyComponent].Pop() : new T();
         }
-        public static void AddToReuseList(short componentKey, IComponent component)
+        public static void AddToReuseList(int componentKey, IComponent component)
         {
             if (_listComponentsReuse.ContainsKey(componentKey))
             {
