+++ DOPFramework/Assets/Dotity/Component/ComponentKey.cs
@@ -1,4 +1,4 @@
-﻿public enum ComponentKey : short
+﻿public enum ComponentKey : int
 {
     Position = 0,
     Speed,
